requirements.txt
pandas
pytest
streamlit
sentence-transformers
faiss-cpu
numpy
spotipy
python-dotenv
